old-scenarios/Assistants/bfsi-bot-in-a-box/requirements.txt
azure-identity==1.14.1
azure-keyvault==4.2.0

openai==1.12.0

numpy==1.26.1
azure-search==1.0.0b2
azure-search-documents==11.4.0

Flask-Cors==4.0.0
Flask==3.0.0

pandas==1.5.2
azure-storage-blob==12.16.0

bs4==0.0.2

python-dotenv==1.0.1